requirements.txt
elasticsearch
Flask
gunicorn
matplotlib
pandas
pefile
PyExifTool
python-magic-win64
recommonmark
scikit-learn
seaborn
signify
Sphinx
tensorflow
yara-python
